further folder refactorings
